src/devika_tokintel/requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0
psutil>=5.9.0

# LLM and AI dependencies
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
sentencepiece>=0.1.99
protobuf>=3.20.0

# Web interface
streamlit>=1.28.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
tqdm>=4.64.0
colorama>=0.4.6
rich>=13.0.0

# Optional: For enhanced features
# uncomment if needed
# accelerate>=0.20.0
# bitsandbytes>=0.41.0
# scipy>=1.10.0
# matplotlib>=3.7.0
# seaborn>=0.12.0 